chore(j-s): Defender Notifications (#11414)
* Improves defender email notifications

* Updates tests

---------
